requirements.txt
langchain==0.0.166
cohere==4.4.1
python-dotenv==1.0.0
transformers==4.29.1
TensorFlow==2.12.0
